perf: Reduce buffering during mixed fast field scans (#2887)
# Ticket(s) Closed

- Closes #2715
- Closes #2623

## What

Adjust the `MixedFastField` executor to stream results, and only buffer
up to a hard coded batch size in memory.

Additionally, avoid creating the intermediate tuple-like `FieldValues`
structure, and directly consume the column values to produce a tuple in
a postgres `Slot`.

## Why

Currently, the `MixedFastField` and `StringFastField` executors buffer
entire columns in memory: that implementation will not scale to larger
datasets. Moving to streaming allows them to be used with arbitrarily
large datasets.

Removing the intermediate `FieldValues` creation also makes
`MixedFastField` ~equivalent to `StringFastField` in terms of overhead:
a followup change will remove `StringFastField`.

#2623 discussed potentially continuing to buffer entire columns and then
declaring them sorted: but doing so efficiently (i.e., without buffering
the entire column) would require a completely different implementation
of the method which started by consuming the dictionary, and then
executed a series of range queries for manageable chunks.

## Tests

Performance improves by ~15% on some of our join queries for the `docs`
dataset.