feat: new SQL builder functions (#2897)
## What

This adds a new set of SQL query "builder functions" that are all
designed to be used with the `@@@` operator such that the field to be
queried must be on the left-hand-side and one of these builder functions
on the right-hand-side.

These functions live in a new schema named `pdb`. We decided to put
these new functions into a new schema for both technical and practical
reasons. The technical reason is to avoid ambiguities when calling them,
as all the original `paradedb.*` builder functions _still_ exist too,
and because `pdb` is easier to type and read than `paradedb` in large
SQL queries.

They all parallel their (pre-existing) counterparts in the `paradedb`
schema


For example, you may have previously written:

```sql
SELECT * FROM mock_items
WHERE id @@@ paradedb.term('description', 'shoes');
```

Now, you can instead write this:

```sql
SELECT * FROM mock_items
WHERE description @@@ pdb.term('shoes');
```

Not only do we hope that this will be easier for humans writing SQL by
hand but for machines generating SQL (think ORM tools). Putting the
field reference to the left of the operator, as opposed to a quoted
string as a function argument is more in-line with typical SQL query
patterns.

---

The full set of new functions in the `pdb` schema encompasses 68
functions (too many to list here) and none of them take a
`paradedb.FieldName` as their first argument. That said, for each
function, there is an existing `paradedb` function counterpart that
does. Not only does this ensure backwards compatibility, but the new
functions essentially rewrite to the old functions during query planning
and/or execution.

For example, this means that the query plans for both the old style of
`WHERE id @@@ paradedb.exists('rating')` and the new, preferred style of
`WHERE rating @@@ pdb.exists()` will be identical, including the json
representation of the "Tantivy Query":

```sql
[v16.2][2212069] test=# explain select * from mock_items where id @@@ paradedb.exists('rating');
                                           QUERY PLAN                                            
-------------------------------------------------------------------------------------------------
 Gather  (cost=1010.00..1014.20 rows=41 width=118)
   Workers Planned: 1
   ->  Parallel Custom Scan (ParadeDB Scan) on mock_items  (cost=10.00..10.10 rows=20 width=118)
         Table: mock_items
         Index: idxmock_items
         Segment Count: 1
         Exec Method: NormalScanExecState
         Scores: false
         Tantivy Query: {"with_index":{"query":{"exists":{"field":"rating"}}}}
(9 rows)

[v16.2][2212069] test=# explain select * from mock_items where rating @@@ pdb.exists();
                                           QUERY PLAN                                            
-------------------------------------------------------------------------------------------------
 Gather  (cost=1010.00..1014.20 rows=41 width=118)
   Workers Planned: 1
   ->  Parallel Custom Scan (ParadeDB Scan) on mock_items  (cost=10.00..10.10 rows=20 width=118)
         Table: mock_items
         Index: idxmock_items
         Segment Count: 1
         Exec Method: NormalScanExecState
         Scores: false
         Tantivy Query: {"with_index":{"query":{"exists":{"field":"rating"}}}}
(9 rows)
```

### Unanticipated Change

It is no longer possible to search for a term "across all text fields".
Previously this was possible via `WHERE id @@@
paradedb.term(value=>'shoes')`, but is no longer supported.

## Why

As part of our development roadmap we're working on making our SQL UX
easier for users (and machines!) in an effort to be more intuitive to
those already familiar with SQL.

## How

By creating a lot of new functions. Technically, these new functions are
the old functions and the old functions are now generated via a custom
proc-macro.

## Tests

All existing unit and regression tests pass.  

There were 3 (?) instances where the "search for a term across all text
fields" was being tested, and those have been commented out (I suspect
we'll have to bring that feature back at some point).

I have yet to implement tests for these new functions. The fact they're
programmatically created makes that a little challenging.

## Docs

I believe the plan is for @rebasedming to work on docs for this in
parallel with docs on the recently added operators (&&&, |||, ###, ===)

---------

Signed-off-by: Eric Ridge <eebbrr@gmail.com>
Co-authored-by: Stu Hood <stuhood@paradedb.com>