chore: Enable running CI for drafts. (#2913)
## What

Enable CI on PR drafts.

## Why

It is good to be able to run CI _before_ notifying your teammates that
something is reviewable. Because we additionally use
required-reviewers/owners, it is currently impossible to run CI without
beginning to email all reviewers.

Moving forward: we should usually only post something as a Draft if we
would like CI to run: for simply sharing a branch with someone, or for
self-review without CI, we can use the
[compare](https://github.com/paradedb/paradedb/compare) view. And we
should only actually mark something reviewable when we would like the
mentioned users to review it.