feat:  "proximity search" support (#2685)
## What

This implements the concept of "proximity search" in pg_search. It has a
lot of overlap in spirit with Lucene's "span" queries.

We support finding a token within some distance (in order, or not) of
another token a set of tokens within some distance of a token or a set
of tokens, and finally one of these proximity clauses within some
distance of a token, a set of tokens, or another proximity clause

Introduces a new datatype named `pdb.proximityclause`, which represents
the complex structure of a proximity clause -- the "left", "distance",
and "right" properties.

Adds a set of new builder functions in the `pdb` schema:

- `pdb.prox_term(term TEXT)`: generates a single term query to be used
as part of a `pdb.proximityclause`
- `pdb.prox_regex(pattern TEXT)`: generates a regular expression query
to be used as part of a `pdb.proximityclause`
- `pdb.prox_clause`, `pdb.prox_clause_in_order`: forms a
`pdb.proximityclause`
- `pdb.prox_array(VARIADIC terms pdb.proximityclause[])`: supports a
variadic array of the above, including itself

And a new overloaded operator `##(pdb.proximityclause, integer)` and
`##(pdb.proximityclause, pdb.proximityclause)` to make writing proximity
clauses via SQL more fluid:

```sql
--
-- find all rows where the body matches the term "bbq" within 3 tokens (in order) of "chicken"
---
SELECT *
FROM t
WHERE body @@@ 'bbq' ##3## 'chicken';
```

## Why

Incredibly valuable text search capability. Essentially it's a form of
fuzzy phrase searching, but with more flexibility.

## How

Continued expansion on our SQL query rewriting capabilities.

## Tests

Regression tests to assert search result correctness along with json
representation

---------

Signed-off-by: Eric Ridge <eebbrr@gmail.com>
Co-authored-by: Ming <ming.ying.nyc@gmail.com>
Co-authored-by: Stu Hood <stuhood@paradedb.com>