perf: Remove StringFastField exec method (#2901)
# Ticket(s) Closed

- Closes #2620

## What

Following up on #2887: remove the `StringFastField` execution method,
since it is never faster than `Mixed`.

Additionally: resolve #2620 by removing the planning conditional around
string aggregate cardinality.

## Why

Simplified code, faster performance.

## How

The string agg cardinality check was occasionally preventing parallelism
on smaller datasets, which turned out to be a good thing: you only
really need enough parallelism to search the segments that you estimate
will be sufficient to answer your query.

The replacement here is an addition to `compute_nworkers` to explicitly
limit the number of workers to the number of segments we think that
we'll need to search (when sorting is not in use but a limit is). See
new comments in that method.

Additionally: this fixes our computation of `nworkers` to account for
the fact that the leader also takes a segment: we were always requesting
at least one worker if we had one segment... when in reality we need
`segments - 1` workers due to the leader. This shifts a lot of explain
plans in regression and unit tests.

## Tests

As shown in
https://github.com/paradedb/paradedb/pull/2887#issuecomment-3116181511,
microbenchmarks are unaffected.

The `docs` dataset has a 1.6x speedup for `line_items-distinct` and ~15%
for a few other queries. The `logs` dataset has 1.35x speedups for
`bucket-string-nofilter` and `bucket-string-filter`.