feat: ability to boost query predicate scores via `::boost(N)` (#2918)
## What

This adds the ability to boost the score of a query clause using
Postgres SQL syntax for operators and casting.

For example:

```sql
SELECT * 
FROM mock_items 
WHERE description @@@ 'shoes'::boost(3);

SELECT * 
FROM mock_items 
WHERE description @@@ pdb.phrase('running shoes')::boost(0.5);
```

It works with the right-hand-side of all our operators: `@@@`, `&&&`,
`|||`, `###`, and `===`, along with the proximity search syntax of
`field @@@ ('one' ##3## 'two')::boost(42)`.

The new `boost` type supports a typmod that gets converted to a `f16`
(so that it can fit within the non-negative side of an i32, which
Postgres requires) and is quietly clamped to the range
`-2048.0..=2028.0`. Any `pdb.query` can be cast to a `boost(N)` where
`N` is either an integer or a floating point.

## Why

Continuation of the SQL UX work to make writing queries easier for
humans and machines.

## How

## Tests

Existing tests pass and a new `boost.sql` regression test has been added
to assert expected outputs.

---------

Signed-off-by: Eric Ridge <eebbrr@gmail.com>