fix: Node relabeling no longer defeats new operators
The new &&&,|||,###,=== operators are all defined to take a value of
type TEXT on the left-hand-side.

If the field used happens to be coereceable to text, but not actually
TEXT, such as VARCHAR, Postgres will wrap that Var node in a
RelabelType, which we need to be able to see through in order to find
the actual field being referenced on the left-hand-side of the operator.