fix: Node relabeling no longer defeats new operators (#2925)
# Ticket(s) Closed

- Closes #

## What

The new &&&,|||,###,=== operators are all defined to take a value of
type TEXT on the left-hand-side.

If the lhs Var happens to be coercible to text, but not actually TEXT,
such as VARCHAR, Postgres will wrap that Var node in a RelabelType,
which we need to be able to see through in order to find the actual
field being referenced on the left-hand-side of the operator.

## Why

## How

## Tests

The existing `operators.sql` regression test has been updated to also
test using a VARCHAR field. I can't think of any other types that
Postgres will relabel to TEXT, especially among the set of data types we
support.