docs: Fix broken Slack link (#2926)
# Ticket(s) Closed

- Closes #N/A

## What
A user reported that the link in our docs was outdated. Somehow all
other links for Slack were this one, but the one in the docs was
different. This fixes that.

## Why
People can join our Slack!

## How
Just used the same invite link as we do everywhere else.

## Tests
I verified it's correct in incognito.