fix: relation extension cache invalidation (#2927) (#2934)
(This ports the recent v0.16.5 changes
(214c9c7dd37a0ddb934e1ce4cbb15f949e1f8a69) forward to `main`)

When extending a relation the backend needs to clear the relation's
`SMgrRelation`'s "SIZE_CACHE" so that it doesn't become confused about
the size of a relation relative to other concurrent relation extensions
that may have occurred.

Failure to do this can cause errors like the below under high read/write
concurrency:

```
ERROR:  XX001: could not read blocks 10..10 in file "base/16384/16552": read only 0 of 8192 bytes
```

PR #2716 introduced this bug as it changed our approach of always
calling `pg_sys::relation_open()` to the new `PgSearchRelation` which
wraps an already-opened `pg_sys::Relation` pointer and is cheaply
clone-able.

Essentially, prior to #2716 we'd always get a new `SMgrRelation` and it
would ask the kernel about the size of the relation on disk, whereas

Fixing this necessitates calling the various
`pg_sys::ExtendBufferedRel*()` functions with the
`pg_sys::ExtendBufferedFlags::EB_CLEAR_SIZE_CACHE` flag set, which also
means we need to use `pg_sys::ExtendedBufferedRel` directly when
extending the relation by one block. So `BM25BufferCache` has been
refactored a bit to handle this.

It's also necessary, when extending the relation by a single buffer, to
lock it using an `ExclusiveLock`, not an `AccessExclusiveLock`.

As a drive-by, this PR adjusts `SegmentComponentWriter`'s flush/drop
behavior to be less confusing and better aided by the Rust compiler.
This is related to the new `LInkedBytesListWriter::finalize_and_write()`
function (see below).

The cleanup around flush & drop also ensures that we won't try to write
any bit of a SegmentComponentWriter's buffers to disk if we're dropping
during a panic-induced stack unwind.

`LinkedBytesListWriter` now has a `fn finalize_and_write(self)` which is
where it records the `last_blockno` in the list's metadata and also
where its `BlockList` is written to disk. The `last_blockno` was
previously being constantly updated by `LinkedBytesListWriter::write()`
every time it linked a new buffer to the end. This wasn't necessarily
incorrect, but it was inefficient and made analyzing the issues this PR
aims to fix a bit more difficult.

Moving the final assignment of `last_blockno` to `finalize_and_write()`
is fine as if the writer is never finalized for whatever reason, the
"last block number" won't matter anyways.

There's a new feature called `block_tracker` that when enabled will
transiently track all block numbers being opened/released and panic when
it detects a block is about to be opened a second time in an
incompatible manner with an already-open instance. This is for internal
debugging and clearly not meant for production use, which is why the
feature is not included in the default feature flag set.

# Ticket(s) Closed

- Closes #

## What

## Why

## How

## Tests

---------

Signed-off-by: Eric Ridge <eebbrr@gmail.com>
Co-authored-by: Stu Hood <stuhood@paradedb.com>