ci: Add missing code from SchemaBot Workflow (#2943)
# Ticket(s) Closed

- Closes #N/A

## What
While doing a quick audit of the CI, I noticed this code is in our
enterprise version, which is running well, and isn't in our community
version. Somehow the community SchemaBot runs aren't even displaying in
the Actions tab.

## Why
I think this code should be there (?)

## How
Compare community and enterprise

## Tests
CI