ci: Update the SSH deploy key to a unified one in organization-wide secrets (#2946)
# Ticket(s) Closed

- Closes #N/A

## What
I had some free time today. I had it on my backlog to unify the deploy
key to not depend on Eric's, and put a universal one across our
organization's secrets instead of a per-repo one, for ease of reuse over
time. This is what this does.

## Why
Simpler cross-repo pushing infra.

## How
^

## Tests
CI