fix: the `LinkedListData::npages` field is dead code
This field was once used to track the number of pages contained in one
of our linked lists, but it's no longer used.

We had a few code paths that updated it (one, incorrectly!) but the
field was never read.

We leave the 4 bytes it occuppied in the `LinkedListData` struct so our
layout remains backwards compatabile, but we no longer use it.