fix: bugs with 0.17.x exposed by the custom FSM
This PR starts off fixing three distinct bugs that we've had for awhile but the new custom FSM brought to light:

- failed to properly identify an orphaned ".deletes" files, which could
  lead to adding the same block to the FSM multiple times
- `BufferMut.return_to_fsm()` needs to unlock/release the buffer before giving the block number to the FSM
- `AtomicGuard::commit()` was returning a block to the FSM prematurely

We also retool the FSM so that it will only return blocks when they're
known to be all visible by all concurrent transactions.  Here on
community that is actually "immediately" but is more sophsicated on our
-enterprise product.

This also allows us to completely remove the concept of the "segment_meta_garbage" list.  Now that the FSM ensures blocks don't get reused until they're allowed, we don't need a separate way of tracking that.

The FSM is also largely rewritten to be centered around a "draining
iterator".

Another minor change is that if we forget to call `AtomicGuard::commit()` its Drop impl now panics