chore: merge 0.17.x to main (fix bugs with 0.17.x exposed by the custom FSM (#2971)) (#2979)
This PR starts off fixing three distinct bugs that we've had for awhile
but the new custom FSM brought to light:

- failed to properly identify an orphaned ".deletes" files, which could
lead to adding the same block to the FSM multiple times
- `BufferMut.return_to_fsm()` needs to unlock/release the buffer before
giving the block number to the FSM
- `AtomicGuard::commit()` was returning a block to the FSM prematurely

We also retool the FSM so that it will only return blocks when they're
known to be all visible by all concurrent transactions. Here on
community that is actually "immediately" but is more sophsicated on our
-enterprise product.

This also allows us to completely remove the concept of the
"segment_meta_garbage" list. Now that the FSM ensures blocks don't get
reused until they're allowed, we don't need a separate way of tracking
that.

The FSM is also largely rewritten to be centered around a "draining
iterator".

Another minor change is that if we forget to call
`AtomicGuard::commit()` its Drop impl now panics

---

And one follow-up change is porting a fix we apparently only made to
-enterprise related to parallel build workers not being able to get the
current transaction id.

That's necessary here too.

---

The full set of individual commits squashed here are:


https://github.com/paradedb/paradedb/pull/2979/commits/95d7b1594c53600a543a8f2340b61fd060d60687

https://github.com/paradedb/paradedb/pull/2979/commits/01d69b2a2a8728ee00190134e852621ee0318c77

https://github.com/paradedb/paradedb/pull/2979/commits/ab7bc08ae59afa239bb87f77d7d27113351d73f5

https://github.com/paradedb/paradedb/pull/2979/commits/7c4ddcd2e315af4d7df00b5386846c66944f7521

https://github.com/paradedb/paradedb/pull/2979/commits/885295995a921682849cc27e412c5c2c7ddf78c4

https://github.com/paradedb/paradedb/pull/2979/commits/a520c5988dc5cc87859d3046e10877c44b93d974

https://github.com/paradedb/paradedb/pull/2979/commits/cd0df327689fa92ba9198f3f92c4df8810108c66

https://github.com/paradedb/paradedb/pull/2979/commits/3f8baf8a4104344bb0f823c01036bc698926415a

https://github.com/paradedb/paradedb/pull/2979/commits/707c55b0a36223c016d33a5e6db16abdbc9a93c6

https://github.com/paradedb/paradedb/pull/2979/commits/c5ee9b5617b07d66cb770f6a088268b7e7d542ad

---------

Co-authored-by: Ming <ming.ying.nyc@gmail.com>
Co-authored-by: Stu Hood <stuhood@paradedb.com>
Co-authored-by: Philippe Noël <21990816+philippemnoel@users.noreply.github.com>