perf: improve FSM extend and pop/pop_many performance
Reduce general overhead in the FSM between extending it and popping off one or more blocks.  In doing so, fix a bug where the header's `empty` flag was not being set as eagerly as it could have been.

Additionally, teach our lower-level new block code to use `RBM_ZERO_AND_LOCK` for when opening buffers returned from the FSM. This potentially avoids disk I/O if the buffer isn't in Postgres' buffer cache and also elides a rountrip over FFI to lock the buffer since ReadBufferExtended is doing it for us with this flag set.

Also cleans up lifetime annotations a bit so it's not possible to have multiple mutable borrows from a BufferMut, even when those borrows happen through Page/PageMut.