chore: Add workflows to automatically cherry-pick from `main` to stable branches (#2998)
## What

Add Github workflows which use
https://github.com/marketplace/actions/github-cherry-pick-action to
automatically open cherry-picks to stable branches for appropriately
labeled PRs.

## Why

This eliminates some annoying manual/local steps when a fix is landing
on `main` which needs to be propagated to stable branches.

## How

To automatically open cherry-pick PRs to `0.17.x` and/or `0.18.x`, use
the `cherry-pick/0.17.x` and `cherry-pick/0.18.x` labels, respectively.