fix: parallel vacuum process now runs correctly (#2987) (#3000)
# Ticket(s) Closed

- Closes #2986

## What

This fixes issue #2986. The primary problem here is that parallel vacuum
processes can't create a transaction id via
`pg_sys::GetCurrentTransactionId()` so we instead teach the FSM to use
`pg_sys::GetCurrentTransactionIdIfAny()` and in the case of a parallel
vacuum process we'll fall back to the
`pg_sys::FirstNormalTransactionId`.

A subtle difference is here that parallel vacuum processes won't be able
to reuse blocks from the FSM in the case where `ambulkdelete()` adds a
".deletes" file to a segment. This is better than not being able to
vacuum at all.

Thanks to @rebasedming for getting a test case together.

Additionally, that test case uncovered a race condition with the
background merge process because it turns out it didn't hold a relation
lock on the index at all. Now it holds a `pg_sys::AccessShareLock`,
which is sufficient for the work it's doing and to prevent concurrent
DDL like `DROP TABLE` from changing things out from under it.

## Why

Fixing bugs.

## How

## Tests

A new regression test has been added and some local stressgres tests
look good.

---------

# Ticket(s) Closed

- Closes #

## What

## Why

## How

## Tests

Co-authored-by: Ming <ming.ying.nyc@gmail.com>