fix: don't ERROR if `paradedb.terms_with_operator` is missing
It's possible that when upgrading pg_search from `v0.15.26` to `v0.17.x`
we could generate an ERROR complaining that
`paradedb.terms_with_operator(...)` doesn't exist if the user upgraded
the extension binary but has not yet run `ALTER EXTENSION pg_search
UPDATE;`.

This fixes that by inspecting the catalogs directly when it's necessary
to lookup that function and plumbing through an `Option<pg_sys::Oid>`
return value.