chore: Open cherry-pick PRs with an automation token. (#3014)
## What

Switch to using the RELEASE token to run cherry-picks.

## Why

Currently the cherry-pick workflow can open PRs, but the tests will not
actually run [because we're using the
`GITHUB_TOKEN`](https://docs.github.com/en/actions/concepts/security/github_token#when-github_token-triggers-workflow-runs).

Ideally this would be a dedicated "automation user", but we'll defer
that for now.