fix: fsm: we have space if any slots are invalid, no need for all (#3015)
## What

When walking the free space map looking for blocks with open slots, we
want to pick a block with any empty slots, not all open slots, so that
we don't end up with a long, low occupancy list.

## Why

Performance.

## How

Add a function to query whether any slots in the fsm block are empty,
and then use it.

## Tests

Ran unit tests.

Co-authored-by: Ori Bernstein <ori@paradedb.com>