fix: Only mark Docker images as `latest` if they actually are the latest ones (#3023)
# Ticket(s) Closed

- Closes #N/A

## What
This PR fixes two issues:

- We were unconditionally marking any non-beta release Docker images as
being `latest`. This means despite having released `0.18.0` before, now
releasing `0.17.7` marked it as latest. This PR fixes this by comparing
the semver versions and creating a flag based on the output, which is
used in the Docker Tag step. This is mostly the same code as in the
`publish-github-release.yml`, where we had a similar issue before.

- I also added some more logs to our `Test ParadeDB Docker` job to help
verify why some users reported permission errors. It seems that the PR
#2954 simply wasn't porteed over to the `0.17.x` release branch.

## Why
Properly tag Docker releases.

## How
^

## Tests
All tested manually.