feat: improved join performance when score and snippets are not used and `@@@` cannot be pushed down to custom scan (#2871)
# Ticket(s) Closed

- Closes #2807

## What

Improved join performance when score and snippets are not used and `@@@`
cannot be pushed down to custom scan

## Why

A source of our JOIN performance degradation is that we’re pushing
`joininfo` (i.e., JOIN-level predicates in
[here](https://github.com/paradedb/paradedb/blob/801e88e9a966b465dee73a381b3902c1a173c5bb/pg_search/src/postgres/customscan/builders/custom_path.rs#L274-L275)
and
[here](https://github.com/paradedb/paradedb/blob/6464dc167c77f043bc01a684f8282467ee464cbf/pg_search/src/postgres/customscan/pdbscan/mod.rs#L264))
down to tantivy, to be able to produce scores and snippets. However,
this almost always leads to a full index scan, which is expected to
produce poor performance. As a quick follow-up, we should either:
1) at least limit this down to the cases that the query is using
snippets or scores, instead of always pushing down the JOIN quals that
almost certainly lead to a full index scan
2) or we can even get stricter and give up on scores and snippets in
this case, too. Then, this will be fixed as soon as CustomScan Join API
is implemented.

In this PR, we go for (1), as it won't have any impact on query results,
and will only improve the performance.

## How

We fallback to index scan (as opposed to custom scan) if the join quals
cannot be pushed down to custom scan, and `score` and `snippet` are not
used in the query.

## Tests

All current tests pass.