perf: optimize merging heuristics to prefer background merging (#3031)
We change the decisions around when/how/what to merge such that we
prefer to merge in the background if we can. That is defined as all of
these being true:

- the user has either a default or non-zero configuration for
`background_layer_sizes`
- no other merge (foreground or background) is currently happening
- the layer policy simulation indicates there is merging work to do 
 
 or

- the request to merge came from a VACUUM

Otherwise we'll merge in the foreground if:

- the request to merge came from `aminsert`
- the user has either a default or non-zero configuration for
`layer_sizes`

Additionally, when merging in the background we now consider the
combined/unique set of layer sizes from both `layer_sizes` and
`background_layer_sizes`, and ensure the maximum layer size is clamped
to our estimate of what an ideal segment size would be based on the
current index size and the `target_segment_count`.

This still allows concurrent backends to merge into layers defined in
`layer_sizes` in the foreground, but they'll prefer to launch that work
in the background if it looks possible.

It does not necessarily ensure there'll only ever be one background
merger per index at any given time, but in practice it's pretty close.