fix: make IndexLayerSizes calculations more accurate (#3039)
## What

The `From<&PgSearchRelation> for IndexLayerSizes` `from()` ctor was
improperly calculating both the total `index_byte_size` and the
`target_segment_byte_size`.

For the former, it needs to only consider SegmentMetaEntries that are
actually visible. In almost every case, at least after a CREATE INDEX of
a large index, the index could actually have 2x+ total segments as there
are actually visible segments.

Secondly, we need to adjust the `target_segment_byte_size` down by 1/3
to account for how `LayeredMergePolicy` does the same thing. Failure to
do this can cause the background worker to merge segments together that
don't actually need to be.

## Why

These two miscalculations are what contributed to the recent query
benchmark regressions we saw after #3031 was merged.

## How

## Tests

Co-authored-by: Eric Ridge <eebbrr@gmail.com>