ci: Fix out-of-order benchmark workflow steps and broken Slack alerts (#3047)
# Ticket(s) Closed

- Closes #N/A

## What
This PR fixes a few issues I randomly noticed:

- Slack alerts from our custom Actions did not trigger. I moved them to
the workflows rather than to the actions, which also enabled cleaning up
a bit of the code.
- Both our Stressgres and Benchmarks jobs had out-of-order steps like
`commit_info`, which meant the output was not available before it was
being called. This must've failed silently until now.
- The `benchmark-backfill-auto.yml` has never run and has been broken.
Eric and I agreed on deleting it, so I've looped that in too.

## Why
Working CI

## How
^

## Tests
N/A