fix: don't include target_segment_count in background_layer_sizes (#3073)
## What

This removes the (sneaky) addition of an "average layer size" based on
the configured `target_segment_count` into the list of
`background_layer_sizes`.

The reason we're removing this is that when this happens on an existing,
large, unbalanced index, it will cause essentially the entire index to
be merged in the background, and that's almost never what one would want
in a production setting.

## Why

Looking to improve operational problems.

## How

## Tests