fix: fix failing test
Our recent changes to tantivy have adjusted the default MemoryArena
page size from 1MB to 512k, and this actually has a downstream effect of
allowing indexers to fit more docs in the arena before tantivy calcuates
that the indexer has exceeded its memory budget.

We have a test that exposed this, so this adjustes the number of
documents to insert to generate 5 segments.