chore: add a label to automated cherry picks (#3107)
## What

In #3095 I removed the `[cherry-pick/x.y.z]` labels from our automated
cherry pick PRs, but after seeing that in action I think some kind of
label will be helpful. So this adds `[automated-cherry-pick]`, which is
not a bright, burning red but a cool, neutral gray.

## Why

To make looking at the PR list a little easier

## How

## Tests