feat: EXPLAIN metrics for parallel workers (#3077)
## What

Record claimed segments and total (re)query counts in the parallel
worker state, and render it in `EXPLAIN ANALYZE VERBOSE`.

```
Parallel Custom Scan (ParadeDB Scan) on public.benchmark_logs  (cost=10.00..10.12 rows=12 width=221) (actual time=128.030..128.174 rows=88 loops=8)
   Output: id, message, country, severity, "timestamp", metadata
   Table: benchmark_logs
   Index: benchmark_logs_idx
   Segment Count: 8
   Heap Fetches: 1
   Virtual Tuples: 0
   Invisible Tuples: 0
   Parallel Workers: {"-1":{"query_count":1,"claimed_segments":["fb215663"]},"0":{"query_count":1,"claimed_segments":["3fa330ca"]},"1":{"query_count":1,"claimed_segments":["a86ca5f4"]},"2":{"query_count":1,"claimed_segments":["bf4ce4c8"]},"3":{"query_count":1,"claimed_segments":["aba90988"]},"4":{"query_count":1,"claimed_segments":["814b1c11"]},"5":{"query_count":1,"claimed_segments":["0d810284"]},"6":{"query_count":1,"claimed_segments":["5ac67c9e"]}}
   Exec Method: TopNScanExecState
   Scores: false
      TopN Order By: id asc
      TopN Limit: 100
      Queries: 8
   Full Index Scan: true
   Tantivy Query: {"with_index":{"query":"all"}}
   Worker 0:  actual time=115.632..115.850 rows=100 loops=1
   Worker 1:  actual time=117.287..117.461 rows=100 loops=1
   Worker 2:  actual time=145.469..145.583 rows=100 loops=1
   Worker 3:  actual time=123.740..123.936 rows=100 loops=1
   Worker 4:  actual time=124.626..124.780 rows=100 loops=1
   Worker 5:  actual time=147.075..147.200 rows=100 loops=1
   Worker 6:  actual time=143.026..143.182 rows=100 loops=1
```

## Why

`EXPLAIN` only runs in the parallel worker leader, and so by default
will not render any metrics recorded by the parallel workers.

We suspect that we have lopsided distributions of segments to workers
and/or re-queries happening in parallel workers, but we don't currently
have visibility into it.

## How

Adding additional fields highlighted the fact that our access to `u32`
fields was not aligned, and that our `transmute`s were resulting in
slices of the wrong length (a `transmute` of a `&[u8]` to a slice of any
other type will result in a slice of the original length!). Instead,
switched to use `std::alloc::Layout` and the `bytemuck` crate to do
these casts more safely.

---------

Co-authored-by: Ming <ming.ying.nyc@gmail.com>