chore: don't warn about `raw` tokenizer on UUID key fields (#3166)
## What

Remove the warning about using the `raw` tokenizer when the `key_field`
is a UUID field.

The drama here is that we (pg_search) assign the `raw` tokenizer to UUID
fields used as the key_field so there's nothing a user can do about it.
Warning about our own bad decisions is not cool.

## Why

Many user and customer complaints.

## How

## Tests

Existing tests pass but a couple of the regression test expected output
is now different.