fix: Clone an Arc rather than a OnceLock. (#3185)
## What

Invert our use of `OnceLock` to ensure that we clone an
`Arc<OnceLock<T>>` rather than a `OnceLock<Arc<T>>`.

## Why

`OnceLock` implements `Clone` by cloning its contents to create a
separate disconnected copy. If what is desired is "exactly once
behavior", then cloning the `OnceLock` before it has been computed the
first time will defeat that.

This change has no impact on benchmarks in this case, but
`Arc<OnceLock<T>>` matches the intent of this code, and sets a better
example for future us.

Co-authored-by: Eric Ridge <eebbrr@gmail.com>