fix: allow user to set TEXT `key_field`'s tokenizer to `keyword` (#3197)
Closes: #2502

## What

This allows the user to change the key_field's tokenizer to `keyword`
since until now we complain that `raw` is deprecated.

## Why

It's not right for us to complain that the `raw` tokenizer is deprecated
and then not allow the user to change it.

In the case of key_field TEXT columns we long long ago made the wrong
decision to use the `raw` tokenizer as the default tokenizer. Then we
decided to deprecate the `raw` tokenizer in favor of the `keyword`
tokenizer, but we never actually let users change the key_field's
tokenizer. So now we do... for TEXT fields only.

## How

## Tests

A new regression test has been added.  All other tests pass.