fix: `paradedb.snippet` no longer empty over partitioned tables (#3204)
# Ticket(s) Closed

- Closes #

## What

In our custom scan we assume that a `Var` belongs to the current range
table entry if the `varno` equals the current range table index.

However, in a partitioned setup, the `varno` can refer to the parent
table while the range table index can refer to the child. This leads to
us not pulling out the `Var`s necessary for snippets to be generated.

## Why

Bug fix

## How

## Tests

Added regression test