fix: Move to using `table_index_fetch_tuple` rather than `heap_fetch`. (#3219)
# Ticket(s) Closed

- Closes #3099

## What

Switch from using `heap_fetch` to using `table_index_fetch_tuple`, and
reuse `TupleTableSlot`s across lookups.

## Why

Two different locations in the codebase were using `heap_fetch`
directly, but `heap_fetch` will not follow HOT updates or deal with
pruned pages.

In the case of snippets, that caused #3099, and in the case of our
heap-filter query, it caused another user reported issue.

## Tests

Added new tests which fail for each of the cases before the fix. The
`regex-and-heap` benchmark runs 20% faster.