chore: Improve determinism of property tests. (#3220)
## What

Set a seed to control what is generated by `random()` in the property
tests, and render it in the reproduction script after a failure.

## Why

To make it easier to reproduce property test failures by running over
reproducible data.

`proptest` failures are only directly reproducible via their reported
seed if they are not dependent on the randomly generated data in the
table that we test against. We can't re-generate the table and index for
every `proptest` query that we run, because it would take way too long
to run a reasonable number of iterations. And we don't want to run on
static data, because then we might never catch data-dependent bugs like
#3266.

Setting a seed allows us to run on random data, but still reproduce
failures later. And for cases where failures aren't data-dependent, the
`proptest` repro seed (e.g. `cc 08176a8c0ae10938a...`) can still be used
directly.