feat: A Free Space Manager fronted by an AVL tree (#3252)
## What

This implements a new `v2` FSM that's fronted by an AVL tree which
allows for minimal locking during extension and draining. It also
provides efficient continuation during drain as xid blocklists are
exhausted or found to be unavailable to the current transaction. And it
implements a (simple) transparent conversion of the current `v1` FSM to
the new format.

Additionally, this fixes a problem with background merging where more
than one background merger process could be spawned at once -- I've seen
up to 8 concurrently. It does this by introducing some a new page on
disk to track the process and coordinate locking.

## Why

Our current FSM is very heavyweight in terms of lock contention. This
should get us to something that isn't.

## How

## Tests

A number of new tests for the array-backed AVL tree and the FSM itself.
All existing tests also pass and, at least, the `wide-table.toml`
stressgres shows a slight performance improvement for the update jobs.

---------

Signed-off-by: Eric Ridge <eebbrr@gmail.com>
Co-authored-by: Stu Hood <stuhood@paradedb.com>