chore: Remove default `255` byte "remove long" filter setting (#3288)
# Ticket(s) Closed

- Closes #

## What

A 255 byte "remove long" setting was applied to all text fields.

This was wrong for text fast fields -- any text longer than 255 bytes
was being silently dropped. And it's confusing to document for non fast
text fields, we should be preserving the tokens and let the user opt
into it rather than doing it by default.

## Why

## How

## Tests