chore: rename `pdb.exact` to `pdb.literal` (#3318)
## What

Renames the new `pdb.exact` tokenizer to `pdb.literal`.

## Why

I think `literal` makes a lot more sense and is a generally more common
way to describe a static value and this is our only opportunity to
change it before we release.

## How

The power of search and replace

## Tests

A number of regression tests and their outputs needed to be adjusted