perf: Introduce a mutable segment type (#3203)
## What

Introduce an optional mutable segment type containing only a list of
ctids, which are then indexed at read time.

## Why

ParadeDB excels at batch inserts, but for single-row INSERT/UPDATE
statements, we currently create single-document Tantivy segments.

These tiny segments have a few problems:

- They consist of up to 8 "files" which are represented as
`LinkedBytesList`s: the minimum overhead of each file is two Postgres
blocks (one for the content, and one for a random-access index). Many of
those blocks will be mostly-empty for a single document segment.
- They increase write amplification. For example: starting from a 1kb
segment means that it takes two merges/compactions to get to 100kb,
representing a 3x write amplification to write 100kb of data.
- They require very rapid updates to the `SegmentMetaEntry`s list: all
inserts insert a new entry, and frequent inserts mean more frequent
merges which also mutate that list.
- Tantivy indexing is relatively fast, but is especially optimized to
operate as a batch process: a single document index technically doesn't
actually need many of the structures that it contains (there is no need
for a DocId because all storage could be boolean, etc).

## How

A mutable segment is read-time indexed when it is loaded (in `mvcc.rs`),
and up to `mutable_segment_rows` rows are inserted into each mutable
segment. Mutable segments are disabled by default, but
`mutable_segment_rows=1000` results in no change to Top-N performance on
the `wide-table` benchmark, and three times faster insert / update
performance.

`SegmentMetaEntry` is bincoded to be stored in the `LinkedItemList` of
entries: to introduce a second variant, we convert a field which became
`SegmentMetaEntry::_unused` in version `0.15.18` into a tag. The`
_unused` field [was previously used to store an xmin
value](http://github.com/paradedb/paradedb/pull/2487). We treat previous
legal values of that tag as representing the legacy encoding.

The insert path defaults to allowing up to `mutable_segment_rows`
inserts to be made into a mutable segment. If more than
`mutable_segment_rows` inserts arrive in the same `aminsert`, we switch
to creating immutable segments for the remainder of that `aminsert`.