chore: Remove `.cursor` folder (#3327)
# Ticket(s) Closed

- Closes #N/A

## What
This ultimately did not gain widespread adoption within the team. I'd
prefer we don't modify the repo for individual workflows, so I'd like to
remove the `.cursor` folder. It's been gitignored so that it can still
be used locally for engineers who wish to without being worried about
pushing it upstream.

## Why
Keep the repo clean of any individual workflow preferences.

## How
N/A

## Tests
N/A